Removed Parithon.StreamDeck.SDK.Core project Moved all code into the SDK project from the Core project.
