+++ Parithon.StreamDeck.SDK/AssemblyStreamDeckApplicationToMonitorAttribute.cs
