+++ Parithon.StreamDeck.SDK/AssemblyStreamDeckAttribute.cs
