+++ Parithon.StreamDeck.SDK/AssemblyStreamDeckOSAttribute.cs
