+++ Parithon.StreamDeck.SDK/Events/KeyUpEvent.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using Newtonsoft.Json;
+using Parithon.StreamDeck.SDK.Models;
 
 namespace Parithon.StreamDeck.SDK.Events
 {
