+++ Parithon.StreamDeck.SDK/Models/Alignment.cs
@@ -1,7 +1,7 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace Parithon.StreamDeck.SDK
+namespace Parithon.StreamDeck.SDK.Models
 {
   [JsonConverter(typeof(StringEnumConverter))]
   public enum Alignment
