+++ Parithon.StreamDeck.SDK/Models/Coordinates.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using Newtonsoft.Json;
 
-namespace Parithon.StreamDeck.SDK
+namespace Parithon.StreamDeck.SDK.Models
 {
   public class Coordinates
   {
