+++ Parithon.StreamDeck.SDK/Parithon.StreamDeck.SDK.csproj
@@ -23,8 +23,4 @@
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="6.0.0" />
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Parithon.StreamDeck.SDK.Core\Parithon.StreamDeck.SDK.Core.csproj" />
-  </ItemGroup>
-
 </Project>
