+++ Parithon.StreamDeck.SDK/StreamDeckAction.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
+using Parithon.StreamDeck.SDK.Messages;
 using Parithon.StreamDeck.SDK.Models;
 
 namespace Parithon.StreamDeck.SDK
