+++ build.ps1
@@ -5,7 +5,6 @@ param
 )
 $Configuration = $(if ($Release.IsPresent) { "Release" } else { "Debug" })
 Write-Output "Building in ${Configuration} mode."
-dotnet build .\Parithon.StreamDeck.SDK.Core --configuration $Configuration
 dotnet build .\Parithon.StreamDeck.SDK.MSBuild.Tool --configuration $Configuration
 dotnet build .\Parithon.StreamDeck.SDK --configuration $Configuration
 dotnet build .\Parithon.StreamDeck.SDK.MSBuild --configuration $Configuration
