+++ sdsdk.sln
@@ -15,8 +15,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		version.json = version.json
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Parithon.StreamDeck.SDK.Core", "Parithon.StreamDeck.SDK.Core\Parithon.StreamDeck.SDK.Core.csproj", "{1D01C134-8E61-4867-B9DD-E84B3524A82A}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Parithon.StreamDeck.SDK.MSBuild", "Parithon.StreamDeck.SDK.MSBuild\Parithon.StreamDeck.SDK.MSBuild.csproj", "{7AC9470C-5CC3-4C97-ABE3-F8B6557BB41B}"
 	ProjectSection(ProjectDependencies) = postProject
 		{4AFA9955-2929-46C9-8FB3-8B8114B5F3B9} = {4AFA9955-2929-46C9-8FB3-8B8114B5F3B9}
@@ -29,7 +27,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Parithon.StreamDeck.SDK.TestConsole", "Parithon.StreamDeck.SDK.TestConsole\Parithon.StreamDeck.SDK.TestConsole.csproj", "{C620F124-180C-4BB8-BE98-4A6EC4D92499}"
 	ProjectSection(ProjectDependencies) = postProject
 		{7AC9470C-5CC3-4C97-ABE3-F8B6557BB41B} = {7AC9470C-5CC3-4C97-ABE3-F8B6557BB41B}
-		{1D01C134-8E61-4867-B9DD-E84B3524A82A} = {1D01C134-8E61-4867-B9DD-E84B3524A82A}
 		{54C7F5E3-78EE-476B-A7D9-25AFDA85D893} = {54C7F5E3-78EE-476B-A7D9-25AFDA85D893}
 	EndProjectSection
 EndProject
@@ -45,10 +42,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1D01C134-8E61-4867-B9DD-E84B3524A82A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D01C134-8E61-4867-B9DD-E84B3524A82A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D01C134-8E61-4867-B9DD-E84B3524A82A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D01C134-8E61-4867-B9DD-E84B3524A82A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{7AC9470C-5CC3-4C97-ABE3-F8B6557BB41B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7AC9470C-5CC3-4C97-ABE3-F8B6557BB41B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7AC9470C-5CC3-4C97-ABE3-F8B6557BB41B}.Release|Any CPU.ActiveCfg = Release|Any CPU
